refactor(feature_assistant): inject app-owned services via AssistantHostAdapter
The moved assistant trees still reached into app/lib for auth, speech,
dictionary, bug reporting, routes, AI preferences, platform detection, and
the Coins ledger. Replace those imports with a single AssistantHostAdapter
seam plus package-owned AssistantRouteNames, and implement the seam once in
app/lib/core/assistant.

The AI preferences settings UI moves to the host as a widget slot: its
settings model, storage dashboard, and model-manager screens are all
app-owned, so mirroring them into the package would duplicate the model to
gain nothing. Its test moves with it. The finance-ingestion integration
test moves to app/ where it can drive the real adapter.
