+++ packages/feature_assistant/test/agent_chat/presentation/screens/model_library_setup_dialog_test.dart
@@ -1,3 +1,4 @@
+import 'package:feature_assistant/src/host/assistant_host_adapter.dart';
 import 'dart:async';
 
 import 'package:feature_assistant/src/agent_chat/application/assistant_model_preferences.dart';
@@ -8,6 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import '../../../support/fake_assistant_host_adapter.dart';
 
 /// Regression cover for the setup dialog that gates the local chat runtime.
 ///
@@ -66,6 +68,9 @@ void main() {
       await tester.pumpWidget(
         ProviderScope(
           overrides: [
+            assistantHostAdapterProvider.overrideWithValue(
+              FakeAssistantHostAdapter(),
+            ),
             assistantModelLibraryProvider.overrideWith((ref) async => state),
             selectedAssistantModelIdProvider.overrideWith(
               (ref) => _SelectedAssistantModelNotifier(),
@@ -140,6 +145,9 @@ void main() {
     await tester.pumpWidget(
       ProviderScope(
         overrides: [
+          assistantHostAdapterProvider.overrideWithValue(
+            FakeAssistantHostAdapter(),
+          ),
           assistantModelLibraryProvider.overrideWith((ref) async => state),
           selectedAssistantModelIdProvider.overrideWith(
             (ref) => _SelectedAssistantModelNotifier(),
