+++ packages/feature_assistant/test/wellbeing/presentation/screens/wellbeing_screen_test.dart
@@ -1,15 +1,24 @@
+import 'package:feature_assistant/src/host/assistant_host_adapter.dart';
 import 'package:feature_assistant/src/wellbeing/presentation/screens/wellbeing_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+import '../../../support/fake_assistant_host_adapter.dart';
 
 /// The wellbeing half of the old Mind hub.
 ///
 /// What matters here is the split itself: this screen keeps reflection and
 /// breathing and must not have brought the AI lab along with it.
 void main() {
-  Widget wrap() =>
-      const ProviderScope(child: MaterialApp(home: WellbeingScreen()));
+  Widget wrap() => ProviderScope(
+    overrides: [
+      assistantHostAdapterProvider.overrideWithValue(
+        FakeAssistantHostAdapter(),
+      ),
+    ],
+    child: const MaterialApp(home: WellbeingScreen()),
+  );
 
   testWidgets('keeps the three wellbeing actions', (tester) async {
     await tester.pumpWidget(wrap());
